replace the wallet1 with a daedalus generated wallet
add recovering in the wallet to show it words
